Initial Working Commit
I've gotten nmma to run with a few warnings using Anaconda and WSL and ran the example files provided by Brandon for GW170817. I have yet to alter any of these files.

Note: the .pkl files are too large to be able to commit to GitHub, so the code will fail if one just clones the repository.